MSDK-124 chat widget sync (PART 1)

Approved-by: Abhishek Sharma <abhishek.sharma@redspace.com>
Approved-by: Shane Kerr <shane.kerr@redspace.com>

+++ livelikesdk/src/main/java/com/livelike/livelikesdk/LiveLikeSDK.kt
@@ -1,7 +1,6 @@
 package com.livelike.livelikesdk
 
 import com.livelike.livelikesdk.messaging.EpochTime
-import java.util.Date
 
 /**
  * Use this class to initialize the LiveLike SDK. This is the entry point for SDK usage. This creates a singleton instance of LiveLike SDK.
@@ -24,7 +23,8 @@ class LiveLikeSDK(appId: String) {
      *  @param currentPlayheadTime
      */
     fun createContentSession(contentId: String,
-                             currentPlayheadTime: ((Long) -> EpochTime)): LiveLikeContentSessionImpl {
+                             currentPlayheadTime: (() -> EpochTime)
+    ): LiveLikeContentSessionImpl {
         return LiveLikeContentSessionImpl(contentId, currentPlayheadTime)
     }
 }
